move settings cards
