+++ apps/nextjs/src/app/settings/profile/page.tsx
@@ -1,11 +1,11 @@
 import { auth } from "@acme/auth";
 import { redirect } from "next/navigation";
 
-import { AvatarCard } from "~/components/settings/profile/avatar-card";
-import { DeleteAccountCard } from "~/components/settings/profile/delete-account-card";
-import { LinksCard } from "~/components/settings/profile/links-card";
-import { TaglineCard } from "~/components/settings/profile/tagline-card";
-import { UsernameCard } from "~/components/settings/profile/username-card";
+import { AvatarCard } from "~/app/settings/profile/avatar-card";
+import { DeleteAccountCard } from "~/app/settings/profile/delete-account-card";
+import { LinksCard } from "~/app/settings/profile/links-card";
+import { TaglineCard } from "~/app/settings/profile/tagline-card";
+import { UsernameCard } from "~/app/settings/profile/username-card";
 import { SettingsSidebar } from "~/components/settings/settings-sidebar";
 
 export default async function ProfileSettingsPage() {
