+++ apps/nextjs/src/app/settings/profile/username-card.tsx
@@ -2,7 +2,7 @@
 import { useState } from "react";
 import { useForm } from "react-hook-form";
 import { useSession } from "next-auth/react";
-import { Button } from "../../ui/Button"
+import { Button } from "../../../components/ui/Button"
 import { api } from "~/trpc/react";
 import {
     type EditProfileInput, editProfileSchema
